+++ Google.Solutions.IapDesktop.Application/Windows/DebugWindow.cs
@@ -21,6 +21,7 @@
 
 using Google.Apis.Auth.OAuth2.Responses;
 using Google.Solutions.IapDesktop.Application.ObjectModel;
+using Google.Solutions.IapDesktop.Application.Services;
 using Google.Solutions.IapDesktop.Application.Settings;
 using Google.Solutions.IapDesktop.Application.Windows.RemoteDesktop;
 using Google.Solutions.IapDesktop.Windows;
